+++ intercom_flutter/ios/Classes/IntercomFlutterPlugin.m
@@ -204,11 +204,18 @@ id windowDidHide;
             NSData *encodedToken=[self createDataWithHexString:token];
             // NSData* encodedToken=[token dataUsingEncoding:NSUTF8StringEncoding];
             NSLog(@"%@", encodedToken);
-            [Intercom setDeviceToken:encodedToken failure:^(NSError * _Nonnull error) {
+            [Intercom setDeviceToken:encodedToken success:^{
+                // Handle success
+                result(@"Token set");
+            } failure:^(NSError * _Nonnull error) {
                 // Handle failure
-                NSLog(@"Error setting device token: %@", error.localizedDescription);
+                NSInteger errorCode = error.code;
+                NSString *errorMsg = error.localizedDescription;
+                
+                result([FlutterError errorWithCode:[@(errorCode) stringValue]
+                                           message:errorMsg
+                                           details: [self getIntercomError:errorCode:errorMsg]]);
             }];
-            result(@"Token set");
         }
     } else if([@"displayArticle" isEqualToString:call.method]) {
         NSString *articleId = call.arguments[@"articleId"];
@@ -299,6 +306,19 @@ id windowDidHide;
                                            details: [self getIntercomError:errorCode:errorMsg]]);
             }];
         }
+    } else if([@"setThemeMode" isEqualToString:call.method]) {
+        NSString *theme = call.arguments[@"theme"];
+        
+        if([@"dark" isEqualToString:theme]){
+            [Intercom setThemeOverride:ICMThemeOverrideDark];
+        } else if([@"light" isEqualToString:theme]){
+            [Intercom setThemeOverride:ICMThemeOverrideLight];
+        } else if([@"system" isEqualToString:theme]){
+            [Intercom setThemeOverride:ICMThemeOverrideSystem];
+        } else {
+            [Intercom setThemeOverride:ICMThemeOverrideNone];
+        }
+        result(@"Theme overridden");
     }
     else {
         result(FlutterMethodNotImplemented);