+++ intercom_flutter_platform_interface/lib/enumeral.dart
@@ -1,0 +1,18 @@
+enum IntercomVisibility {
+  gone,
+  visible,
+}
+
+enum IntercomTheme {
+  // // Enable dark mode
+  dark,
+
+  // Enable light mode
+  light,
+
+  // Use system preference
+  system,
+
+  // Clear override and use server-provided theme
+  none,
+}